Fix :mojira: emote ID in help command (#55)
When the Mojira logo changed to match the [new Mojang Studios logo](https://www.minecraft.net/en-us/article/meet-mojang-studios) the :mojira: emote was replaced. This caused the internal Discord emote ID to change. Unfortunately, Discord doesn't have a feature to access customs emotes without hardcoding the ID, so this is liable to break if the emote ID ever changes again.